requirements.txt
# HTTP and async
aiohttp==3.9.1
requests==2.31.0

# HTML parsing
beautifulsoup4==4.12.2

# Telegram
python-telegram-bot==20.7

# RSS parsing
feedparser==6.0.11

# Database
aiosqlite==0.19.0

# Configuration
python-dotenv==1.0.0

# Retry logic (keeping for compatibility)
tenacity==8.2.3

# Development and compatibility
dataclasses-json==0.6.3